Location, Room and Building rewrited

+++ app/Http/Controllers/BuildingController.php
@@ -24,9 +24,9 @@ class BuildingController extends Controller
         ])->validate();
 
         if ($validator and Building::create($request->all())) {
-            LogController::log('Gebäude erstellt', '{"name": "' . $request->name . '", "location_id": "' . $request->location_id . '"}');
+            // LogController::log('Gebäude erstellt', '{"name": "' . $request->name . '", "location_id": "' . $request->location_id . '"}');
 
-            return redirect()->back()->with('success', __('Msg.LocationCreated'));
+            return redirect()->back()->with('success', __('Msg.BuildingCreated'));
         }
         return redirect()->back()->withErrors(['message' => 'Building could not be created']);
     }
@@ -41,13 +41,15 @@ class BuildingController extends Controller
     public function update(UpdateBuildingRequest $request, Building $building)
     {
         $validator = Validator::make($request->all(), [
-            'name' => 'required|unique:buildings|max:100',
+            'name' => 'required|max:100',
+            'id' => 'required|integer|exists:buildings,id',
+            'location_id' => 'required|integer',
         ])->validate();
 
         if ($validator and $building->whereId($request->input('id'))->update($request->except(['_token', '_method']))) {
-            LogController::log('Gebäude aktualisiert', '{"name": "' . $request->name . '"}');
+            // LogController::log('Gebäude aktualisiert', '{"name": "' . $request->name . '"}');
 
-            return redirect()->back()->with('success', __('Msg.LocationUpdated'));
+            return redirect()->back()->with('success', __('Msg.BuildingUpdated'));
         }
         return redirect()->back()->withErrors(['message' => 'Building could not be updated']);
     }
@@ -60,12 +62,22 @@ class BuildingController extends Controller
      */
     public function destroy(Request $building)
     {
-        $find = Building::find($building->input('id'));
-        if ($find->delete()) {
-            LogController::log('Gebäude gelöscht', '{"name": "' . $building->name . '"}');
+        $validator = Validator::make($building->all(), [
+            'id' => 'required|integer|exists:buildings,id',
+        ])->validate();
 
-            return redirect()->back()->with('success', __('Msg.LocationDeleted'));
+        $find = Building::find($building->input('id'));
+
+        if($find->rooms()->count() > 0) {
+            return redirect()->back()->withErrors(['message' => 'Building could not be deleted, because it has rooms assigned to it.']);
         }
+
+        if ($find->delete()) {
+            // LogController::log('Gebäude gelöscht', '{"name": "' . $building->name . '"}');
+
+            return redirect()->back()->with('success', __('Msg.BuildingDeleted'));
+        }
+
         return redirect()->back()->with('message', 'Could not delete building');
     }
 }
